requirements.txt

pydantic-settings
pydantic
loguru
pypdf
python-docx
openpyxl
openai
tenacity
python-dotenv
reportlab
pillow
google-genai
protobuf